ready homework 2

+++ app/Http/Controllers/ProductController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Product;
 use Illuminate\Http\Request;
+use App\Http\Requests\CrearProductRequest;
+use App\Http\Requests\EditProductRequest;
 
 class ProductController extends Controller
 {
@@ -16,7 +18,7 @@ class ProductController extends Controller
     {
         //
         $products = Product::get();
-        return response()->json($products);
+        return response()->json($products, 200);
     }
 
     /**
@@ -35,10 +37,11 @@ class ProductController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(CrearProductRequest $request)
     {
         // Create a new product
         $product = Product::create($request->all());
+        
 
         // Return a response with a product json
         // representation and a 201 status code   
@@ -51,10 +54,19 @@ class ProductController extends Controller
      * @param  \App\Product  $product
      * @return \Illuminate\Http\Response
      */
-    public function show(Product $product)
+    public function show($id)
     {
-        //
-        return response()->json($product, 200);
+        $response = '';
+        $status = 500;
+        $product = Product::find($id);
+        if($product != null) {
+            $response = $product;
+            $status = 200;
+        } else {
+            $response = ['errors' => ['code' => 'ERROR-2', 'title' => 'Not Found']];
+            $status = 404;
+        }
+        return response()->json($response, $status);
     }
 
     /**
@@ -75,15 +87,21 @@ class ProductController extends Controller
      * @param  \App\Product  $product
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Product $product)
+    public function update(EditProductRequest $request, $id)
     {
-        //
-        print($request);
-        
-        if($request->input('name') != null){$product->{'name'} = $request->input('name');}
-        if($request->input('price') != null){$product->{'price'} = $request->input('price');}
-        $product->save();
-        return response()->json($product, 200);
+        $response = '';
+        $status = 500;
+        $product = Product::find($id);
+        if($product != null) {
+            $product->fill($request->all());
+            $product->save();
+            $status = 200;
+            $response = $product;
+        } else {
+            $response = ['errors' => ['code' => 'ERROR-2', 'title' => 'Not Found']];
+            $status = 404;
+        }
+        return response()->json($response, $status);
     }
 
     /**
@@ -92,13 +110,19 @@ class ProductController extends Controller
      * @param  \App\Product  $product
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Product $product)
+    public function destroy($id)
     {
-        //
-        $body = json_decode($product);
-        $id = $body->{'id'};
-        Product::destroy($id);
-        return response()->json(204);
+        $response = '';
+        $status = 500;
+        $product = Product::find($id);
+        if($product != null) {
+            $status = 204;
+            Product::destroy($id);
+        } else {
+            $response = ['errors' => ['code' => 'ERROR-2', 'title' => 'Not Found']];
+            $status = 404;
+        }        
+        return response()->json($response, $status);
 
     }
 }
